Add users api - wip
