+++ beepbeep/dataservice/views/__init__.py
@@ -1,4 +1,5 @@
 from .home import home
 from .swagger import api
+from .users import users_api
 
-blueprints = [home, api]
+blueprints = [home, api, users_api]
